+++ src/test/java/de/alpharogroup/compare/object/CompareObjectExtensionsTest.java
@@ -29,6 +29,7 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -37,6 +38,9 @@ import org.apache.commons.lang3.StringUtils;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
+import com.google.common.collect.Sets;
+
+import de.alpharogroup.test.objects.Permission;
 import de.alpharogroup.test.objects.Person;
 import de.alpharogroup.test.objects.enums.Gender;
 
@@ -77,7 +81,7 @@ public class CompareObjectExtensionsTest
 		// explanation of expected: compared object are equal
 		expected = true;
 		actual = CompareObjectExtensions.compare(sourceOjbect, objectToCompare);
-		assertEquals("Cloned object should be equal with the source object.", expected, actual);
+		assertEquals(expected, actual);
 		// 2. scenario...
 		// expected: the compare method should return false...
 		// explanation of expected: compared object are not equal
@@ -87,9 +91,7 @@ public class CompareObjectExtensionsTest
 
 		expected = false;
 		actual = CompareObjectExtensions.compare(sourceOjbect, objectToCompare);
-		assertEquals(
-			"Object to compare should be not equal with the source object because it has changed.",
-			expected, actual);
+		assertEquals(expected, actual);
 	}
 
 	/**
@@ -183,21 +185,21 @@ public class CompareObjectExtensionsTest
 	{
 		int expected;
 		int actual;
-		Person sourceOjbect;
-		Person objectToCompare;
+		Permission sourceOjbect;
+		Permission objectToCompare;
 		// create a person...
-		sourceOjbect = Person.builder().gender(Gender.MALE).name("obelix").build();
+		sourceOjbect = Permission.builder().name("read").build();
 		// make a clone of it...
-		objectToCompare = Person.builder().gender(Gender.MALE).name("obelix").build();
+		objectToCompare = Permission.builder().name("read").build();
 
 		expected = 0;
 		actual = CompareObjectExtensions.compareTo(sourceOjbect, objectToCompare);
-		assertEquals("Given objects should not be equal.", expected, actual);
+		assertEquals(expected, actual);
 
-		sourceOjbect.setName("asterix");
-		expected = -14;
+		sourceOjbect.setName("write");
+		expected = 5;
 		actual = CompareObjectExtensions.compareTo(sourceOjbect, objectToCompare);
-		assertEquals("Given objects should not be equal.", expected, actual);
+		assertEquals(expected, actual);
 	}
 
 	/**
@@ -223,6 +225,8 @@ public class CompareObjectExtensionsTest
 		Person asterix;
 		Person miraculix;
 		List<Person> persons;
+		Comparator<String> comparator;
+		Comparator defaultComparator;
 
 		persons = new ArrayList<>();
 		obelix = Person.builder().gender(Gender.MALE).name("obelix").build();
@@ -231,18 +235,10 @@ public class CompareObjectExtensionsTest
 		// 1. scenario...
 		actual = CompareObjectExtensions.compareTo(asterix, obelix, "name");
 
-		final Comparator<String> comp = new Comparator<String>()
-		{
+		comparator = Comparator.naturalOrder(); // (o1, o2) -> o1.compareTo(o2);
+		expected = comparator.compare(asterix.getName(), obelix.getName());
 
-			@Override
-			public int compare(final String o1, final String o2)
-			{
-				return o1.compareTo(o2);
-			}
-		};
-		expected = comp.compare(asterix.getName(), obelix.getName());
-
-		assertEquals("Result of compared properties should be equal.", expected, actual);
+		assertEquals(expected, actual);
 
 		persons.add(obelix);
 		persons.add(asterix);
@@ -254,8 +250,8 @@ public class CompareObjectExtensionsTest
 
 		log.log(Level.FINE, "Unsorted Persons:");
 		log.log(Level.FINE, persons.toString());
-		final Comparator defaultComparator = new BeanComparator("name");
-		Collections.sort(persons, defaultComparator);
+		defaultComparator = new BeanComparator("name");
+		persons.sort(defaultComparator);
 
 		log.log(Level.FINE, "Sorted Persons by name:");
 		log.log(Level.FINE, persons.toString());
@@ -271,6 +267,94 @@ public class CompareObjectExtensionsTest
 		assertEquals(persons.get(0), obelix);
 		assertEquals(persons.get(1), miraculix);
 		assertEquals(persons.get(2), asterix);
+	}
+
+	/**
+	 * Test method for {@link CompareObjectExtensions#compareTo(Object, Object, Set)}
+	 *
+	 * @throws IllegalAccessException
+	 *             Thrown if this {@code Method} object is enforcing Java language access control
+	 *             and the underlying method is inaccessible.
+	 * @throws InvocationTargetException
+	 *             Thrown if the property accessor method throws an exception
+	 * @throws NoSuchMethodException
+	 *             Thrown if a matching method is not found or if the name is "&lt;init&gt;"or
+	 *             "&lt;clinit&gt;".
+	 *
+	 */
+	@Test(enabled = true)
+	public void testCompareToSetProperties()
+		throws IllegalAccessException, InvocationTargetException, NoSuchMethodException
+	{
+		int expected;
+		int actual;
+		Permission sourceOjbect;
+		Permission objectToCompare;
+		// create a person...
+		sourceOjbect = Permission.builder().name("read").description("Permission for read files")
+			.build();
+		// make a clone of it...
+		objectToCompare = Permission.builder().name("read").description("Permission for read files")
+			.shortcut("R").build();
+		// compare only name and description
+		expected = 0;
+		actual = CompareObjectExtensions.compareTo(sourceOjbect, objectToCompare,
+			Sets.newHashSet("name", "description"));
+		assertEquals(expected, actual);
+		// compare all
+		expected = -1;
+		actual = CompareObjectExtensions.compareTo(sourceOjbect, objectToCompare,
+			Sets.newHashSet("name", "description", "shortcut"));
+		assertEquals(expected, actual);
+		// alter name and compare only name and description
+		sourceOjbect.setName("write");
+		expected = 5;
+		actual = CompareObjectExtensions.compareTo(sourceOjbect, objectToCompare,
+			Sets.newHashSet("name", "description"));
+		assertEquals(expected, actual);
+		// compare only description
+		expected = 0;
+		actual = CompareObjectExtensions.compareTo(sourceOjbect, objectToCompare,
+			Sets.newHashSet("description"));
+		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link CompareObjectExtensions#compareTo(Object, Object, Set)}
+	 *
+	 * @throws IllegalAccessException
+	 *             Thrown if this {@code Method} object is enforcing Java language access control
+	 *             and the underlying method is inaccessible.
+	 * @throws InvocationTargetException
+	 *             Thrown if the property accessor method throws an exception
+	 * @throws NoSuchMethodException
+	 *             Thrown if a matching method is not found or if the name is "&lt;init&gt;"or
+	 *             "&lt;clinit&gt;".
+	 *
+	 */
+	@Test(enabled = true)
+	public void testCompareToSetPropertiesWithComparableObjects()
+		throws IllegalAccessException, InvocationTargetException, NoSuchMethodException
+	{
+		int expected;
+		int actual;
+		Person sourceOjbect;
+		Person objectToCompare;
+		// create a person...
+		sourceOjbect = Person.builder().gender(Gender.MALE).name("obelix").build();
+		// make a clone of it...
+		objectToCompare = Person.builder().gender(Gender.MALE).name("obelix").build();
+
+		expected = 0;
+		actual = CompareObjectExtensions.compareTo(sourceOjbect, objectToCompare,
+			Sets.newHashSet("gender", "name"));
+		assertEquals(expected, actual);
+
+		sourceOjbect.setName("asterix");
+		expected = -14;
+		actual = CompareObjectExtensions.compareTo(sourceOjbect, objectToCompare,
+			Sets.newHashSet("gender", "name"));
+		assertEquals(expected, actual);
 	}
 
 	/**
@@ -295,11 +379,96 @@ public class CompareObjectExtensionsTest
 		String objectToCompare;
 		// create a person...
 		sourceOjbect = Person.builder().gender(Gender.MALE).name("obelix").build();
-		;
+
 		// make a clone of it...
 		objectToCompare = "";
 
 		CompareObjectExtensions.compareTo(sourceOjbect, objectToCompare);
+	}
+
+	/**
+	 * Test method for {@link CompareObjectExtensions#compare(Object, Object, String...)}
+	 *
+	 * @throws IllegalAccessException
+	 *             Thrown if this {@code Method} object is enforcing Java language access control
+	 *             and the underlying method is inaccessible.
+	 * @throws InvocationTargetException
+	 *             Thrown if the property accessor method throws an exception
+	 * @throws NoSuchMethodException
+	 *             Thrown if a matching method is not found or if the name is "&lt;init&gt;"or
+	 *             "&lt;clinit&gt;".
+	 *
+	 */
+	@Test(enabled = true)
+	public void testCompareToVarargsProperties()
+		throws IllegalAccessException, InvocationTargetException, NoSuchMethodException
+	{
+		int expected;
+		int actual;
+		Permission sourceOjbect;
+		Permission objectToCompare;
+		// create a person...
+		sourceOjbect = Permission.builder().name("read").description("Permission for read files")
+			.build();
+		// make a clone of it...
+		objectToCompare = Permission.builder().name("read").description("Permission for read files")
+			.shortcut("R").build();
+		// compare only name and description
+		expected = 0;
+		actual = CompareObjectExtensions.compare(sourceOjbect, objectToCompare, "name",
+			"description");
+		assertEquals(expected, actual);
+		// compare all
+		expected = -1;
+		actual = CompareObjectExtensions.compare(sourceOjbect, objectToCompare, "name",
+			"description", "shortcut");
+		assertEquals(expected, actual);
+		// alter name and compare only name and description
+		sourceOjbect.setName("write");
+		expected = 5;
+		actual = CompareObjectExtensions.compare(sourceOjbect, objectToCompare, "name",
+			"description");
+		assertEquals(expected, actual);
+		// compare only description
+		expected = 0;
+		actual = CompareObjectExtensions.compare(sourceOjbect, objectToCompare, "description");
+		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link CompareObjectExtensions#compareTo(Object, Object)}
+	 *
+	 * @throws IllegalAccessException
+	 *             Thrown if this {@code Method} object is enforcing Java language access control
+	 *             and the underlying method is inaccessible.
+	 * @throws InvocationTargetException
+	 *             Thrown if the property accessor method throws an exception
+	 * @throws NoSuchMethodException
+	 *             Thrown if a matching method is not found or if the name is "&lt;init&gt;"or
+	 *             "&lt;clinit&gt;".
+	 *
+	 */
+	@Test(enabled = true)
+	public void testCompareToWithComparableObjects()
+		throws IllegalAccessException, InvocationTargetException, NoSuchMethodException
+	{
+		int expected;
+		int actual;
+		Person sourceOjbect;
+		Person objectToCompare;
+		// create a person...
+		sourceOjbect = Person.builder().gender(Gender.MALE).name("obelix").build();
+		// make a clone of it...
+		objectToCompare = Person.builder().gender(Gender.MALE).name("obelix").build();
+
+		expected = 0;
+		actual = CompareObjectExtensions.compareTo(sourceOjbect, objectToCompare);
+		assertEquals(expected, actual);
+
+		sourceOjbect.setName("asterix");
+		expected = -14;
+		actual = CompareObjectExtensions.compareTo(sourceOjbect, objectToCompare);
+		assertEquals(expected, actual);
 	}
 
 	/**
@@ -326,7 +495,7 @@ public class CompareObjectExtensionsTest
 		Person objectToCompare;
 		// create a person...
 		sourceOjbect = Person.builder().gender(Gender.MALE).name("obelix").build();
-		;
+
 		// make a clone of it...
 		objectToCompare = null;
 
@@ -451,7 +620,7 @@ public class CompareObjectExtensionsTest
 		expected = 5;
 		actual = compareToResult.size();
 
-		assertEquals("size of map should be 5 but is " + actual + ".", expected, actual);
+		assertEquals(expected, actual);
 	}
 
 	/**